Inform websocket client task to tear down websocket client
This prevents a race condition between the EspWebSocketClient
instance and the websocket client task.
If EspWebSocketClient's drop() destroys the websocket client handle,
it can destroy this resource while the websocket client task is
using it. Specifically, the websocket client task checks the
selected_for_destroying flag and will conditionally destroy the
client handle itself. This can lead to a double free.